feat: Refactor complete for API hooks
