+++ package.json
@@ -17,7 +17,7 @@
     "@mysten/sui": "^1.11.0",
     "@mysten/zklogin": "^0.7.21",
     "@reduxjs/toolkit": "^2.2.7",
-    "@tanstack/react-query": "^5.56.2",
+    "@tanstack/react-query": "^5.62.2",
     "axios": "^1.7.7",
     "bignumber.js": "^9.1.2",
     "clsx": "^2.1.1",
@@ -31,7 +31,6 @@
     "react-helmet": "^6.1.0",
     "react-hot-toast": "^2.4.1",
     "react-icons": "^5.3.0",
-    "react-query": "^3.39.3",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.26.2",
     "react-select": "^5.8.1"
