+++ src/api/hooks/auth/index.js
@@ -1,5 +1,5 @@
+import { post } from "../.."
 import { useMutation } from "@tanstack/react-query"
-import { post } from "../../../api"
 
 export const useCreateAccountMutation = () => {
   return useMutation({
