+++ src/api/hooks/mail/index.js
@@ -1,15 +1,14 @@
 import { useMutation } from "@tanstack/react-query"
-// import { sendMail, deleteMail } from "../../../services/mail.service"
 import { deleteMail, sendMail } from "../../../services/mail.service"
 
 export const useSendMaillQuery = () => {
   return useMutation({
-    mutationFn: () => sendMail(),
+    mutationFn: (sendMailDto) => sendMail(sendMailDto),
   })
 }
 
 export const useDeleteMaillQuery = () => {
   return useMutation({
-    mutationFn: () => deleteMail(),
+    mutationFn: (email) => deleteMail(email),
   })
 }
