+++ src/views/auth/login/index.jsx
@@ -2,7 +2,7 @@ import { useState } from "react"
 import { toast } from "react-hot-toast"
 import { Link } from "react-router-dom"
 import { isValidEmail } from "../../../utils"
-import { useLoginMutation } from "../../../redux/hooks/auth"
+import { useLoginMutation } from "../../../api/hooks/auth"
 import { AuthService } from "../../../services/auth.service"
 import { PiEyeClosedLight, PiEyeLight } from "react-icons/pi"
 import ConnectWalletModal from "../../../components/modals/ConnectWalletModal"
